LayoutTests/http/tests/media/media-source/video-media-source-seek-expected.txt

Tests seeking with MediaSource API

EVENT(webkitsourceopen)
EVENT(webkitsourceended)
EVENT(playing)
EVENT(timeupdate) : seeking to 2.78
EVENT(seeking)
EVENT(seeked)
EVENT(playing)
EVENT(ended)
END OF TEST
